https://ds1projapailden.shinyapps.io/hiringpredictor/
This application used Random forest method to determine the chances of New hire employees to churn in less than 3 months. 80% of the data is used as the training set while 20% utilized as test set. The impact in accuracy of each variable is also included in the view.

Random forest create multiple decision trees and combined together to get a more accurate and stable prediction It is a popular machine learning algorithm.